We have made extensive changes in this version. We fixed the issue where the Content Browser project would get stuck in renaming when added from outside the editor. We also resolved the problem where rotating multiple entities at once would result in incorrect rendering of the rotation handles. Additionally, we fixed the crash issue that occurred when simulation was stopped while in 2D mode. We removed the stack trace for GetComponent when the component was not found. Furthermore, we addressed the loading lag issue (we now properly support loading from assets/scripts/binary files, as well as project directory/binary files). Lastly, we specifically added support for HTTPS in the version control functionality of the engine.
We have made extensive changes in this version. We fixed the issue where the Content Browser project would get stuck in renaming when added from outside the editor. We also resolved the problem where rotating multiple entities at once would result in incorrect rendering of the rotation handles. Additionally, we fixed the crash issue that occurred when simulation was stopped while in 2D mode. We removed the stack trace for GetComponent when the component was not found. Furthermore, we addressed the loading lag issue (we now properly support loading from assets/scripts/binary files, as well as project directory/binary files). Lastly, we specifically added support for HTTPS in the version control functionality of the engine.